+++ test/stubs/email-service-stub.ts
@@ -1,4 +1,5 @@
-import { IEmailOptions, IEmailService } from '@/use-cases/interfaces';
+import { IEmailOptions, IEmailService } from '@/use-cases/services/email-service/interfaces';
+import { SendEmailService } from '@/use-cases/services/email-service/send-email';
 
 export const makeEmailService = (): IEmailService => {
   class EmailServiceStub implements IEmailService {
@@ -7,4 +8,11 @@ export const makeEmailService = (): IEmailService => {
     }
   }
   return new EmailServiceStub();
+};
+
+export const makeSendEmailService = (): SendEmailService => {
+  const sendEmailService = new SendEmailService(
+    makeEmailService(),
+  );
+  return sendEmailService;
 };
