+++ test/stubs/verification-token-use-cases-stub.ts
@@ -1,7 +1,5 @@
 import { makeVerificationTokenRepository, makeUniversallyUniqueIdentifierGenerator, makeHasher } from '@/test/stubs';
 import { AddVerificationTokenUseCase } from '@/use-cases/verification-token/add-verification-token';
-import { makeEmailService } from '@/test/stubs/email-service-stub';
-import { SendVerificationTokenUseCase } from '@/use-cases/verification-token/send-verification-token';
 
 export const makeAddVerificationTokenUseCase = (): AddVerificationTokenUseCase => {
   const VerificationTokenRepositoryStub = makeVerificationTokenRepository();
@@ -11,11 +9,4 @@ export const makeAddVerificationTokenUseCase = (): AddVerificationTokenUseCase =
     VerificationTokenRepositoryStub, tokenGeneratorStub, hasher,
   );
   return AddVerificationTokenUseCaseStub;
-};
-
-export const makeSendVerificationTokenUseCase = (): SendVerificationTokenUseCase => {
-  const sendVerificationTokenUseCase = new SendVerificationTokenUseCase(
-    'http://url.test', makeEmailService(),
-  );
-  return sendVerificationTokenUseCase;
 };
