+++ test/use-cases/services/email-service/send-email/send-email-service.spec.ts
@@ -1,18 +1,16 @@
 import { makeEmailService } from '@/test/stubs/email-service-stub';
-import { EmailServiceError } from '@/use-cases/errors';
-import { IEmailService } from '@/use-cases/interfaces';
-import { SendVerificationTokenUseCase } from '@/use-cases/verification-token/send-verification-token';
+import { EmailServiceError } from '@/use-cases/services/email-service/errors';
+import { IEmailService } from '@/use-cases/services/email-service/interfaces';
+import { SendEmailService } from '@/use-cases/services/email-service/send-email';
 
 type SutTypes = {
-  sut: SendVerificationTokenUseCase,
+  sut: SendEmailService,
   emailService: IEmailService,
 };
 
 const makeSut = (): SutTypes => {
   const emailService = makeEmailService();
-  const sut = new SendVerificationTokenUseCase(
-    'http://url.test', emailService,
-  );
+  const sut = new SendEmailService(emailService);
 
   return {
     sut,
@@ -20,16 +18,17 @@ const makeSut = (): SutTypes => {
   };
 };
 
-describe('SendVerificationTokenUseCase', () => {
+describe('SendEmailService', () => {
   it('Should call send', async () => {
     const { sut, emailService } = makeSut();
 
     const emailServiceSpy = jest.spyOn(emailService, 'send');
 
     await sut.execute({
-      user: { id: 'any_userId', name: 'any_name', email: 'any_email@test.com' },
-      token: 'any_token',
-      expiresInHours: 1,
+      user: { name: 'any_name', email: 'any_email@test.com' },
+      subject: 'any_subject',
+      text: 'any_text',
+      html: 'any_html',
     });
     expect(emailServiceSpy).toHaveBeenCalledTimes(1);
   });
@@ -40,9 +39,10 @@ describe('SendVerificationTokenUseCase', () => {
     jest.spyOn(emailService, 'send').mockReturnValueOnce(new Promise((_, reject) => reject(new Error())));
 
     const response = await sut.execute({
-      user: { id: 'any_userId', name: 'any_name', email: 'any_email@test.com' },
-      token: 'any_token',
-      expiresInHours: 1,
+      user: { name: 'any_name', email: 'any_email@test.com' },
+      subject: 'any_subject',
+      text: 'any_text',
+      html: 'any_html',
     });
     expect(response.isError()).toBe(true);
     expect(response.value).toEqual(new EmailServiceError());
@@ -54,9 +54,10 @@ describe('SendVerificationTokenUseCase', () => {
     jest.spyOn(emailService, 'send').mockResolvedValue(null);
 
     const response = await sut.execute({
-      user: { id: 'any_userId', name: 'any_name', email: 'any_email@test.com' },
-      token: 'any_token',
-      expiresInHours: 2,
+      user: { name: 'any_name', email: 'any_email@test.com' },
+      subject: 'any_subject',
+      text: 'any_text',
+      html: 'any_html',
     });
     expect(response.isSuccess()).toBe(true);
     expect(response.value).toBe(null);
